+++ src/test/feature-debug.spec.ts
@@ -2,6 +2,7 @@ import { Address, toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
 import { __DANGER_resetNodeId } from "../grammar/ast";
 import { Debug } from "./features/output/debug_Debug";
+import path from "path";
 
 describe("feature-debug", () => {
     beforeEach(() => {
@@ -29,13 +30,24 @@ describe("feature-debug", () => {
             res.indexOf("=== DEBUG LOGS ===") + 19,
             res.indexOf("=== VM LOGS ===") - 2,
         );
-        expect(debugLogs).toStrictEqual(`stack(2 values) : 10000000000 () 
+
+        const filePath = path.normalize("src/test/features/debug.tact");
+
+        expect(debugLogs).toStrictEqual(`[DEBUG] File ${filePath}:10:9
+stack(2 values) : 10000000000 () 
+[DEBUG] File ${filePath}:11:9
 Hello world!
+[DEBUG] File ${filePath}:12:9
 123
+[DEBUG] File ${filePath}:13:9
 true
+[DEBUG] File ${filePath}:14:9
 false
+[DEBUG] File ${filePath}:15:9
 null
+[DEBUG] File ${filePath}:16:9
 ${contract.address.toString({ bounceable: true })}
+[DEBUG] File ${filePath}:17:9
 ${Address.parseRaw(
     "0:83dfd552e63729b472fcbcc8c45ebcc6691702558b68ec7527e1ba403a0f31a8",
 )}`);